+++ test/infrastructure/native/native_bcp_runtime_test.dart
@@ -1,3 +1,4 @@
+import 'package:odbc_fast/domain/entities/driver_capabilities.dart';
 import 'package:odbc_fast/infrastructure/native/native_bcp_runtime.dart';
 import 'package:test/test.dart';